Refactoring: moving FTPServerSession into ServerSession class. Fixing other bugs.
