+++ hw-project/src/main/java/ru/spbau/mit/ftp/Server.java
@@ -4,17 +4,13 @@ import org.apache.commons.cli.*;
 
 import java.io.*;
 import java.net.InetSocketAddress;
-import java.nio.channels.AsynchronousCloseException;
-import java.nio.channels.ServerSocketChannel;
-import java.nio.channels.SocketChannel;
+import java.nio.channels.*;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Scanner;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
-import java.util.concurrent.Future;
 import java.util.concurrent.TimeUnit;
-import java.util.concurrent.atomic.AtomicInteger;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -24,9 +20,8 @@ public class Server extends AbstractServer {
     private static final Logger LOGGER = LogManager.getLogger("server");
     private final int nthreads;
     private ExecutorService executorService;
-    private final List<SocketChannel> sockets = new ArrayList<>();
-    private final List<Future> futures = new ArrayList<>();
     private ServerSocketChannel serverSocketChannel;
+    private List<SocketChannel> sockets = new ArrayList<>();
 
     public Server(int nthreads) {
         this.nthreads = nthreads;
@@ -71,10 +66,14 @@ public class Server extends AbstractServer {
 
     @Override
     public synchronized void start(String hostName, int port) throws IOException, ServerException {
-        executorService = Executors.newFixedThreadPool(nthreads);
 
         if (serverSocketChannel != null && serverSocketChannel.isOpen()) {
             throw new ServerException("server already running");
+        }
+        try {
+            executorService = Executors.newFixedThreadPool(nthreads);
+        } catch (IllegalArgumentException e) {
+            throw new ServerException(e);
         }
 
         serverSocketChannel = ServerSocketChannel.open();
@@ -90,9 +89,9 @@ public class Server extends AbstractServer {
                     }
                     SocketChannel socketChannel = serverSocketChannel.accept();
                     if (socketChannel != null) {
+                        executorService.submit(new ServerSession(socketChannel));
                         synchronized (this) {
                             sockets.add(socketChannel);
-                            futures.add(executorService.submit(new FTPServerSession(socketChannel)));
                         }
                     }
                 }
@@ -111,27 +110,20 @@ public class Server extends AbstractServer {
     public synchronized void stop() throws IOException {
         LOGGER.debug("entering stop");
         if (serverSocketChannel != null) {
-            if (sockets.size() != futures.size()) {
-                LOGGER.fatal("sockets.size() != futures.size()");
-            }
-            for (int i = 0; i < sockets.size(); i++) {
-                SocketChannel socketChannel = sockets.get(i);
-                Future future = futures.get(i);
-                if (socketChannel.isOpen()) {
+            serverSocketChannel.close();
+            executorService.shutdown();
+            synchronized (this) {
+                for (SocketChannel socketChannel : sockets) {
                     socketChannel.close();
                 }
-                future.cancel(true);
+                sockets.clear();
             }
-            sockets.clear();
-            futures.clear();
-            serverSocketChannel.close();
-        }
-        executorService.shutdown();
-        while (!executorService.isTerminated()) {
-            try {
-                executorService.awaitTermination(1, TimeUnit.SECONDS);
-            } catch (InterruptedException ignored) {
-                break;
+            while (!executorService.isTerminated()) {
+                try {
+                    executorService.awaitTermination(1, TimeUnit.SECONDS);
+                } catch (InterruptedException ignored) {
+                    break;
+                }
             }
         }
         LOGGER.debug("exiting stop");
@@ -140,74 +132,5 @@ public class Server extends AbstractServer {
     @Override
     public void close() throws IOException {
         stop();
-    }
-
-    private static class FTPServerSession implements Runnable {
-        private static final Logger LOGGER = LogManager.getLogger("session");
-        private static final AtomicInteger SESSIONS_COUNT = new AtomicInteger(0);
-        private final SocketChannel socketChannel;
-        private final int sessionId;
-
-        FTPServerSession(SocketChannel socketChannel) {
-            this.socketChannel = socketChannel;
-            this.sessionId = SESSIONS_COUNT.incrementAndGet();
-        }
-
-        @Override
-        public void run() {
-            try {
-                EchoResponse.INIT_RESPONSE.write(socketChannel);
-                LOGGER.info(logMessage("starting IO loop"));
-                while (true) {
-                    Request request = Request.parse(socketChannel);
-                    Response response = dealRequest(request);
-                    response.write(socketChannel);
-                    LOGGER.debug(logMessage("sent: " + response + ": " + response.debugString()));
-                    if (response.equals(EchoResponse.EXIT_RESPONSE)) {
-                        break;
-                    }
-                }
-                LOGGER.info(logMessage("closing socket"));
-            } catch (Exception e) {
-                LOGGER.error(logMessage(e.toString()));
-            } finally {
-                try {
-                    if (socketChannel.isOpen()) {
-                        EchoResponse.EXIT_RESPONSE.write(socketChannel);
-                        socketChannel.close();
-                    }
-                } catch (IOException e) {
-                    LOGGER.error(logMessage(e.toString()));
-                }
-            }
-            LOGGER.info(logMessage("exiting..."));
-        }
-
-        private Response dealRequest(Request request) throws ServerException {
-            LOGGER.info(logMessage("received: " + request + ": " + request.debugString()));
-            Response response;
-            if (request instanceof EchoRequest) {
-                String receivedMessage = ((EchoRequest) request).getMessage();
-                if (receivedMessage.equals(EchoRequest.EXIT_MESSAGE)) {
-                    response = EchoResponse.EXIT_RESPONSE;
-                } else {
-                    response = new EchoResponse(receivedMessage);
-                }
-            } else if (request instanceof ListRequest) {
-                String path = ((ListRequest) request).getPath();
-                File[] files = new File(path).listFiles();
-                response = new ListResponse(files);
-            } else if (request instanceof GetRequest) {
-                String path = ((GetRequest) request).getPath();
-                response = GetResponse.serverGetResponse(path);
-            } else {
-                throw new ServerException("Unknown request: " + request);
-            }
-            return response;
-        }
-
-        private String logMessage(String message) {
-            return String.format("#%d: %s", sessionId, message);
-        }
     }
 }
